Getting rid of compiler warnings.
On branch bojan_fixes_test_cases_in_manual

modified:   Convert/Convert_Mod.f90
modified:   Convert/Main_Con.f90
modified:   Convert/makefile_explicit_dependencies
modified:   Divide/makefile_explicit_dependencies
modified:   Generate/makefile_explicit_dependencies
modified:   Process/Cpu/Main_Pro.f90
modified:   Process/Cpu/Process_Mod/Convective_Outflow.f90
modified:   Process/Cpu/Process_Mod/Initialize_Variables.f90
modified:   Process/Cpu/Swarm_Mod.f90
modified:   Process/Cpu/Swarm_Mod/Sgs_Discrete_Random_Walk.f90
modified:   Process/Cpu/makefile
modified:   Process/Cpu/makefile_explicit_dependencies
deleted:    Shared/Control_Mod/Input_Output/Profiler_Info.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Cell_Centers.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Cell_Inertia.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Cell_Volumes.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Face_Centers.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Face_Geometry.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Face_Interpolation.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Face_Surfaces.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Faces_Surface.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Global_Volumes.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Wall_Distance.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Weights_Cells_To_Nodes.f90
deleted:    Shared/Grid_Mod/Calculate/Calculate_Weights_Nodes_To_Cells.f90
deleted:    Shared/Probe_1d_Cells.f90
deleted:    Shared/Probe_1d_Nodes.f90
deleted:    Shared/Probe_1d_Nodes.h90
deleted:    Shared/Work_Mod/Safe/Connect_Int_Cell.f90
deleted:    Shared/Work_Mod/Safe/Connect_Int_Face.f90
deleted:    Shared/Work_Mod/Safe/Connect_Int_Node.f90
deleted:    Shared/Work_Mod/Safe/Connect_Real_Cell.f90
deleted:    Shared/Work_Mod/Safe/Connect_Real_Face.f90
deleted:    Shared/Work_Mod/Safe/Connect_Real_Node.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Int_Cell.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Int_Face.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Int_Node.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Real_Cell.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Real_Face.f90
deleted:    Shared/Work_Mod/Unsafe/Connect_Real_Node.f90
new file:   Convert/Convert_Mod/Insert_Buildings.f90

+++ Sources/Process/Cpu/Swarm_Mod.f90
@@ -69,8 +69,8 @@
     real :: dt  !! time step for the swarm
 
     ! Eddy time interval for stochastic eddy interaction (SEIM) model
-    integer :: time_eim  !! eddy time interval for stochastic
-                         !! eddy interaction (SEIM) model
+    real :: time_eim  !! eddy time interval for stochastic
+                      !! eddy interaction (SEIM) model
     ! Number of sub-steps; time sub-steps
     integer :: n_sub_steps  !! number of sub-steps in time
